8240275: Fix race between GC and KeepAliveCache (HttpURLConnection)
When using HttpURLConnection or HttpsURLConnection, the underlying HttpClient object is normally added to the KeepAliveCache for reuse when the response body has been fully read.  However, if the response body is never fully read (for example, if a 201 response code is returned by the server and the code that is using HttpURLConnection reads the "Location" header but doesn't bother reading the response body), then (prior to this commit) the underlying HttpClient object is added to the KeepAliveCache by the MeteredStream/KeepAliveStream finalizer when the HttpURLConnection/HttpClient/InputStream/etc is garbage collected.

Unfortunately, "Resurrecting" the HttpClient object from the MeteredStream/KeepAliveStream finalizer like this is unreliable and leads to intermittent failures in subsequent HttpURLConnection calls.  For example, GC may identify that both the MeteredStream and Socket underlying an HttpURLConnection are ready to be finalized, and may finalize the MeteredStream before the Socket, causing the HttpClient object to be "resurrected" and added to the KeepAliveCache.  If a new HttpURLConnection is created (in another thread that is running in parallel with GC) after this happens and before GC gets around to finalizing the Socket, the new HttpURLConnection may reuse this "resurrected" HttpClient.  If GC then gets around to finalizing the associated Socket while the new HttpURLConnection is actively using it, the Socket finalizer will close the underlying connection while it is actively being used by the new HttpURLConnection and cause an error.  If sun.net.http.retryPost is set to "false" then this error may propagate up to user-level code.  Alternatively, if the above condition is combined with conditions that trigger the bug described in https://bugs.openjdk.java.net/browse/JDK-8209178 then such errors can lead to hung connections after the HttpClient automatically retries the interrupted connection.

This commit fixes this issue by adding each HttpClient object to a new data structure in KeepAliveCache immediately after it is created, then using a WeakReference/ReferenceQueue on the HttpURLConnection/HttpsURLConnection to detect when the associated HttpURLConnection/HttpsURLConnection is garbage collected so the KeepAliveCache can either drop the HttpClient reference (so the HttpClient can be garbage collected) or move the HttpClient to the "idle and available for reuse" data structure as appropriate.  This eliminates the current "object resurrection" behavior, ensures that references to the HttpClient and all underlying objects are maintained for the entire lifespan of each HttpClient, and prevents GC from finalizing an HttpClient or any underlying objects while they are still in use.